+++ lib/faker/internet.ex
@@ -1,4 +1,6 @@
 defmodule Faker.Internet do
+  import Faker, only: [sampler: 2]
+
   alias Faker.Name.En, as: Name
 
   @moduledoc """